Refactoring
Moved Implemented CLI commands to sub-folder and Helpers to a sub-folder.
