+++ webfiori/framework/cli/commands/CreateCommand.php
@@ -22,10 +22,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-namespace webfiori\framework\cli;
+namespace webfiori\framework\cli\commands;
 
+use webfiori\framework\cli\CLICommand;
 use webfiori\database\Table;
 use webfiori\framework\Util;
+use webfiori\framework\cli\helpers\CreateTable;
+use webfiori\framework\cli\helpers\CreateTableObj;
+use webfiori\framework\cli\helpers\CreateWebService;
+use webfiori\framework\cli\helpers\CreateMiddleware;
+use webfiori\framework\cli\helpers\CreateCLIClassHelper;
+use webfiori\framework\cli\helpers\CreateCronJob;
 /**
  * A command which is used to automate some of the common tasks such as 
  * creating table classes or controllers.
