+++ webfiori/framework/cli/commands/RunSQLQueryCommand.php
@@ -23,8 +23,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace webfiori\framework\cli;
+namespace webfiori\framework\cli\commands;
 
+use webfiori\framework\cli\CLICommand;
 use webfiori\database\DatabaseException;
 use webfiori\database\Table;
 use webfiori\framework\DB;
