+++ webfiori/framework/cli/commands/VersionCommand.php
@@ -23,8 +23,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace webfiori\framework\cli;
+namespace webfiori\framework\cli\commands;
 
+use webfiori\framework\cli\CLICommand;
+use webfiori\framework\cli\CLI;
 /**
  * Description of VersionCommand
  *
