+++ webfiori/framework/cli/helpers/CreateCLIClassHelper.php
@@ -1,6 +1,7 @@
 <?php
-namespace webfiori\framework\cli;
+namespace webfiori\framework\cli\helpers;
 
+use webfiori\framework\cli\commands\CreateCommand;
 /**
  * A helper class which is used to help in creating cron jobs classes using CLI.
  *
